member details code structure revamp password reset integration with backend
